freebsd: Extract ioctl hidraw to separate module
