+++ framework_lib/src/touchpad.rs
@@ -1,78 +1,47 @@
+#[cfg(not(target_os = "freebsd"))]
 use hidapi::{HidApi, HidDevice, HidError};
+
 #[cfg(target_os = "freebsd")]
-use nix::{ioctl_read, ioctl_read_buf, ioctl_readwrite, ioctl_readwrite_buf, ioctl_write_buf};
-#[cfg(target_os = "freebsd")]
-use std::fs::OpenOptions;
-use std::io::{Read, Write};
+use crate::freebsd_hid::*;
 #[cfg(target_os = "freebsd")]
 use std::os::fd::AsRawFd;
-#[cfg(target_os = "freebsd")]
-use std::os::unix::fs::OpenOptionsExt;
-
-#[cfg(target_os = "freebsd")]
-#[repr(C)]
-pub struct HidIocGrInfo {
-    bustype: u32,
-    vendor: u16,
-    product: u16,
-}
-
-#[cfg(target_os = "freebsd")]
-//ioctl_readwrite!(hidraw_get_report_desc, b'U', 21, HidrawGetReportDesc);
-//ioctl_readwrite!(hidraw_get_report, b'U', 23, HidrawGetReport);
-//ioctl_write!(hidraw_set_report, b'U', 24, HidrawSetReport);
-ioctl_read!(hidiocgrawninfo, b'U', 32, HidIocGrInfo);
-//ioctl_readwrite!(hidiocgrawnname, b'U', 33, HidIocGrName);
-ioctl_read_buf!(hid_raw_name, b'U', 33, u8);
-ioctl_write_buf!(hid_set_feature, b'U', 35, u8);
-ioctl_readwrite_buf!(hid_get_feature, b'U', 36, u8);
 
 pub const PIX_VID: u16 = 0x093A;
+pub const TP_PID: u16 = 0x0274;
 pub const PIX_REPORT_ID: u8 = 0x43;
 
 #[cfg(target_os = "freebsd")]
 pub fn print_touchpad_fw_ver() -> Option<()> {
-    println!("Touchpad");
-    let path = "/dev/hidraw2";
-    let mut file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .custom_flags(libc::O_NONBLOCK)
-        .open(path)
-        .unwrap();
+    if let Some(file) = hidraw_open(PIX_VID, TP_PID) {
+        println!("Touchpad");
+        unsafe {
+            let fd = file.as_raw_fd();
 
-    let mut desc = HidIocGrInfo {
-        bustype: 0,
-        vendor: 0,
-        product: 0,
-    };
-    unsafe {
-        let fd = file.as_raw_fd();
-        if let Err(err) = hidiocgrawninfo(fd, &mut desc) {
-            error!("Failed to access hidraw at {}: {:?}", path, err);
-            return None;
+            let mut desc = HidIocGrInfo {
+                bustype: 0,
+                vendor: 0,
+                product: 0,
+            };
+            if let Err(err) = hidiocgrawninfo(fd, &mut desc) {
+                error!("Failed to call hidiocgrawninfo: {}", err);
+                return None;
+            }
+            println!("  IC Type:           {:04X}", desc.product);
+
+            let mut buf = [0u8; 255];
+            if let Err(err) = hid_raw_name(fd, &mut buf) {
+                error!("Failed to call hid_raw_name: {}", err);
+                return None;
+            }
+            let name = std::str::from_utf8(&buf)
+                .unwrap()
+                .trim_end_matches(char::from(0));
+            debug!("  Name: {}", name);
+
+            println!("  Firmware Version: v{:04X}", read_ver(fd)?);
+
+            read_byte(fd, 0x2b);
         }
-        debug!(
-            "Found {:04X}:{:04X} Bustype: {:04X}",
-            desc.vendor, desc.product, desc.bustype
-        );
-        // TODO: Iterate over all /dev/hidraw* devices and find the right one
-        // if vid != PIX_VID || (pid != 0x0274 && pid != 0x0239) {
-        println!("  IC Type:           {:04X}", desc.product);
-
-        let mut buf = [0u8; 255];
-        if let Err(err) = hid_raw_name(fd, &mut buf) {
-            error!("Failed to access hidraw at {}: {:?}", path, err);
-            return None;
-        }
-        let name = std::str::from_utf8(&buf)
-            .unwrap()
-            .trim_end_matches(char::from(0));
-        debug!("  Name: {}", name);
-
-        println!("  Firmware Version: v{:04X}", read_ver(fd)?);
-
-        read_byte(fd, 0x2b);
     }
 
     Some(())
